requirements.txt
Flask==2.0.1
requests==2.25.1
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-auth-httplib2==0.1.0
google-api-python-client==2.12.0
pandas==1.2.4
celery==5.1.2
redis==3.5.3